Refactor connection deletion in pool
Moved the deletion and closing of connections in pool.go from within the loops iterating through idle and full connections to its own loop. This change simplifies the code, prevents repetition, and makes it clear when and how connections are removed from the pool. In addition, it prepares for potential functionality for handling deletion errors or conn absence.