Replace custom atomicCounter with atomic.Uint64
This commit replaces the custom atomicCounter type with the built-in atomic.Uint64 from the sync/atomic package which fit better for usage tracking. In addition, `Used` field is added in the stats struct to keep track of the use count of connection. This change simplifies the code and potentially improves performance by using the more efficient atomic operations provided by the standard library.