Rename atomic.go to ptr.go and update counter type
To make the code more efficient, the custom atomicCounter type has been replaced with the standard atomic.Uint64 from the sync/atomic package. This change was made to utilize standard library for atomic operations and also to introduce a new `Used` field in the stats struct for accurate usage tracking.