Rename statsUnlocked() to statsUnsafe()
In the codebase, statsUnlocked() was renamed to statsUnsafe() to better reflect its functionality. The method assumes the mutex is already locked and should not be directly accessed to get statistics, hence the previous name might have been misleading. The new name, statsUnsafe(), highlights the potential risks associated with incorrect use.

Please make sure to use the public Stats() method to safely acquire statistics. This change was implemented in three places in the pool.go file.