Optimize memory allocation for Stats construction
To optimize memory allocation in `statsUnsafe` function, initial capacity is now set for the `Connections` slice in the `Stats` struct. This change is made considering that the length of `p.connMap` would be the maximum possible length of the `Connections` slice, thus providing an initial allocation that should reduce subsequent reallocations.