Change pool connMap to storage for clarity
Changed all 'connMap' references to 'storage' in the pool.go and pool_test.go files. This renaming to 'storage' provides clearer understanding that this map is storing all the GRPC connections handled by the pool, including those that are to be closed, and not just mapping connections. This naming change helps in enhance code readability and maintainability.