Adjust comment for 'nonBlocking' field in 'options.go'
Updated a comment associated with the 'nonBlocking' field in 'options.go'. The reason for the change is to better communicate that currently there is no method to change this field and to emphasize that in most cases, the blocking option is preferred. This adds clarity for code maintainers and developers, reducing potential confusion or misuse of this field.