Refine mutex comment in 'pool.go'
Improved the comment referring to the mutex variable in the 'pool.go' file. It now clarifies that the mutex is used to protect both read and write access to storage and connections. The enhancement was necessary to better convey how the mutex is used and hence, aid in better code comprehension.