Remove nullLogger struct, replace with LoggerFunc
This commit eliminates the nullLogger struct and replaces it with a LoggerFunc to increase efficiency. Previous implementation of nullLogger was found to be redundant, not serving any impactful purpose. Replacing with LoggerFunc provides the same logging capability with cleaner and more streamlined code. The test for logger is also updated to reflect this change.