Phonkee bugfixes (#3)
* Update section title in README.md for clarity

The "algorithm" section title in README.md was misleading as it gave an impression of explaining the under-the-hood mechanism of code. Instead, this section provides usage examples. Updated the title to "how" to better reflect the contents and improve readability for users.

* Fix error handling in pool.go

Adjusted the error handling in pool.go to ensure that the context error and the ErrMaxConnectionsReached error are wrapped properly. Previously, the ErrMaxConnectionsReached was being passed as a value and not wrapped correctly. This commit corrects that in order to provide more accurate error information.

* "Fix error handling when maximum connections exceeded in pool.go

Swapped the order of 'ctx.Err()' and 'ErrMaxConnectionsReached' in the 'fmt.Errorf' function to correctly wrap the errors. Previously, 'ErrMaxConnectionsReached' was not properly wrapped, which could cause unclear error messages. The purpose of this adjustment is to present more precise error information when the maximum number of connections is reached."

* "Add context to Pool Close method and handle cleanup asynchronously

Adjusted the Close() method in pool.go to include a context parameter. This allows for better control of the close process. An anonymous goroutine was added to perform cleanup of connections in the background and check for no remaining connections in a concurrent-safe manner. A select statement was introduced to allow the method to exit once all connections are closed or when the provided context ends. This could greatly improve the responsiveness of the program when shutting down, especially under heavy load situations."